requirements.txt
fastapi==0.75.0
networkx==2.7.1
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.5
openai==0.15.0
pydantic==1.8.2
gunicorn==20.1.0
uvicorn==0.11.7